core: Handle stale events in Thing.schedule
Currently there is a window between an event firing and calling the
time.AfterFunc and state.Parse being called. During this window the BRL
is not held and it is possible for the Thing of the event to be freed.
In this case Thing.Free cannot cancel the event as it has already fired
resulting in state.Parse being called on a freed Thing.

This commit updates Thing.schedule to acquire the BRL early and check if
the Thing the event is for has already been freed.

  If the Thing is freed the event is stale, the BRL is released and the
  event is not processed. A stale event cannot be logged for event
  debugging as the details would have been lost when the Thing freed.

  If the Thing is not freed the event is manually processed via
  state.parse with the already held BRL.

There is no additional overhead of acquiring the BRL early as it would
be acquired by state.Parse anyway. It does eliminate the overhead of
acquiring and releasing the BRL when event debugging is active.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>