core: Add state.subparseFor to perform commands as alternative actor
Currently it would be very difficult and messy for an actor to cause
another actor to perform a command. For example a GIVE command that is
implemented as an actor performing DROP on and item and having the
receiver perform a GET on the item - thus reusing the existing DROP/GET
command's logic.

This commit adds a simple state.subparseFor method to cause another
actor to perform a command. Existing buffers are used automatically.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>