core: In Thing.Junk/Free cancel all in-flight events
Currently if a Thing is junked and then freed Thing.Free will cancel the
events using Suspend and then deleting the event. If the Thing is junked
but not freed events are left in-flight. This could cause events to fire
for a Thing when it is out of play.

When a Thing is junked ANY in-flight events should always be cancelled.

This commit updates Thing.Junk to cancel in-flight events for non-freed
Thing. The Thing.Free method has been updated to use Cancel instead of
Suspend and delete, as this makes the intent clearer.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>