client: Upgrade players to have health when created/loaded
This commit updates client.assemblePlayer to add a Health record if one
is not found. If an old style (v0.0.15 and above) Health record using
FREQUENCY and REGENERATES is found it will be upgraded so that FREQUENCY
is AFTER and REGENERATES is RESTORE.

The client.createPlayer method has been updated to add default Health
values.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>