core: Add health events to regenerate health (new $HEALTH command)
This commit adds the new $HEALTH command and ties it to the Health
event. The $POOF command has been updated to register the health event
if the player's current health is less than their maximum health. When a
player reaches maximum health a message is displayed:

  H:28/30>
  You feel healthy.
  H:30/30>

Observers will see:

  H:30/30>
  Diddymus looks healthy.
  H:30/30>

Note: The only way to test this currently is to edit a player file, set
current health < maximum health, login and wait.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>