core: Fix NPC/Player flags and flag usage for Thing
Currently a player has the Player and NPC flags set. The NPC flag is
incorrectly set by Thing.Unmarshal as we cannot set the PC flag before
unmarshaling the player.

This commit updates Thing.Untracked to detect an NPC as anything having
HealthMaximum set and body parts. The confusing references to Player in
Thing.Unmarshal have been removed.

The Player flag is set and NPC flag cleared in cliet.assemblePlayer.

The TAKE command has been updated to check for the NPC or Player flag.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>