cmd: Fix TAKE/PUT commands used on players
Currently, if the TAKE or PUT commands are used on a player a confusing
message is seen by the player:

  You see no '#UID-1E9' to take anything from.
  You see no '#UID-1E9' to put anything into.

This is because the player is matched but then not resolved in the
commands because the location Who is not checked.

This commit adds the checks to resolve players at the location for the
TAKE and PUT commands.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>