core,docs: Permit leading '+' to be omitted for bound qualifiers
Currently a non-bound qualifier has a leading '+'. For example:

  Alias: +WOODEN +SHORT:SWORD SHORTSWORD

The non-bound alias is 'WOODEN'. The bound alias is 'SHORT' bound to
'SWORD'. When writing zone files it is very easy to accidentally write
'SHORT:SWORD' and omit the leading '+'. When this happens the qualifier
becomes 'HORT:SWORD' due to the first character being dropped as it is
assumed to be a leading '+'. This can lead to a non-obvious, confusing
bug.

As we can deduce a bound alias from the colon ':' this commit changes
the Thing.Unmarshal method to allow bound qualifiers with or without the
leading '+' - although a '+' is preferred to match non-bound qualifiers.
Thus either of the following is now permitted and have the same effect:

  Alias: +WOODEN +SHORT:SWORD SHORTSWORD
  Alias: +WOODEN SHORT:SWORD SHORTSWORD

A note for this has been added to docs/zone-files.txt

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>