core: Add ANY modifier to matcher
While actions can be scripted to address players the action always
targets the first player in a location's Who list. For example:

  OnAction: HIT PLAYER
          : EXAMINE PLAYER

Here it is always the first player hit and the first player examined.

This commit introduces the ANY modifier to the matcher. ANY will return
one random item from the list of items matched. This allows the
following to be written:

  OnAction: HIT ANY PLAYER
          : EXAMINE ANY PLAYER

This results in a random player being picked for the HIT and EXAMINE
commands.

It should be noted that any suffers the same limitation as other
modifiers. For example:

  GET 1ST BALL 1ST BALL

Results in only the first ball being taken no matter how many there are.
In the same way:

  GET ANY BALL ANY BALL ANY BALL

may return one, two or three balls dependant on wether the same random
ball is pick more than once.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>