core: Added matcher.StripMatch to remove a match from original input
Currently it is vary difficult to remove the words used in a match from
the original player input. For example, if a player enters:

  talk to any city guard The vendor is here!

We can parse the command into "TALK ANY CITY GUARD VENDOR HERE!" and
identify the guard from "ANY CITY GUARD" using the matcher. However, for
the TALK command we need the original input minus the text used to
identify the guard. In this case we want to turn "to any city guard The
vendor is here!" into "The vendor is here!".

A new matcher function StripMatch will remove the stopwords, modifiers,
qualifiers and alias from the original input for a match. For example:

  talk to any city guard The vendor is here!
  ^    ^  ^   ^    ^     ^                 ^
  |    |  |   |    |     |_________________|
  |    |  |   |    alias  remaining message
  |    |  |   qualifier
  cmd  |  modifier
       stop word

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>