core: Remove Thing.Event map entry when event suspended
In commit f9c18388e66f399c: "core: remove Thing.Event map entry when
event cancelled/cleaned-up" a change was made to clean-up Thing.Event
when an event was cancelled.

This should also occur when an event is suspended. Currently the map
entry for a suspended event is set to nil.

This commit moves the Thing.Event map delete from the Thing.cancel
method, added by the referenced commit above, to the Thing.suspend
method replacing setting the map entry to nil. This works for both
cancelled and suspended events due to cancel calling suspend.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>