client: Limit player names to 15 letters max at account creation
Currently there is no limit to the length of a character name.

This commit limits player names to a maximum of 15 characters. The
longer the name the harder it is for people to type it. Maybe there
should be an automatic alias for the first few characters?

Note: Existing players will retain any long names...

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>