core: Maintain a map of in-game players
Currently there is no easy way of finding out who is currently in-game.
The only way would be to iterate through all locations and containers
looking for players :/

This commit adds a state.Players map of UIDs to player Thing. A player
is added to the list by the $POOF command and removed by the QUIT
command. The list of in-game players has a number of uses: for commands
such as /WHO, for locating players, summoning players to a location or
teleporting to a player's location.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>