core,doc: Ignore DOOR attributes in location definitions
Currently a DOOR attribute included as part of a location's definition
results in a nil panic.

Allowing a DOOR to be defined as part of a location raises a number of
issue: Where does the alias to open it come from? What is it's
description? How do we create the 'other side'?

For now a DOOR attribute in a location's definition is ignored.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>