core: Fix body check fail when all slots in use
When items are held, worn or wielded the body slots used are removed
from Any[Body]. When the used item is removed the slots specified by the
items Any[Holdable], Any[Wearable] or Any[Wieldable] are added back to
the Any[Body]. However, if all slots are used Any[Body] becomes empty
and it appears that there is no body. This means that body checks using
len(Any[Body]) will fail.

This commit adds a new isKey HasBody. When Unmarshaling a Thing HasBody
is set if there is BODY attribute. The container check is more precise
in Thing.Unmarshal - any Thing that is a location or has a body is not a
container. The check in the EXAMINE command is more precise and can
check specifically for a body.

The more precise body checks now mean that items can be defined with
bodies, for example see the doll in data/zinara.wrj,  and they are not
miss-identified as NPCs.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>