core: change NPCs to anything that can self heal & regenerate health
Currently NPCs are defined and identified as anything that has maximum
health OR body parts. This prevents certain items from being defined
properly - such as the doll in data/zone/zinara.wrj which has a body but
is not intended to be an NPC.

This commit changes the definition of an NPC to be anything that can
self heal and regenerate health.

These changes will allow items in the future to define HEALTH attributes
but are not themselves incorrectly identified as NPCs. Any combination
of MAXIMUM, CURRENT, RESTORE, AFTER+JITTER can be set, as long as
MAXIMUM, RESTORE and AFTER+JITTER are not set together. It should be
noted that such items can also have negative effects and cause harm.

Some possible examples:

  // A ring of health, adds +5 to maximum health when worn:
  %%
        Ref: O1
       Name: a ring
      Alias: RING
   Wearable: FINGER
     Health: MAXIMUM→5

  This is a small, plain brass ring.
  %%

  // A ring of regeneration, speeds up regeneration by 30s when worn:
  %%
        Ref: O2
       Name: a ring
      Alias: RING
   Wearable: FINGER
     Health: AFTER→30s

  This is a small, plain brass ring.
  %%

  // A ring of healing, increase regeneration by +5 when worn:
  %%
        Ref: O3
       Name: a ring
      Alias: RING
   Wearable: FINGER
     Health: RESTORE→5

  This is a small, plain brass ring.
  %%

  // A ring of weakness, reduces maximum health by -10 when worn:
  %%
        Ref: O3
       Name: a ring
      Alias: RING
   Wearable: FINGER
     Health: RESTORE→5

  This is a small, plain brass ring.
  %%

  // A potion of healing, increases current health by +30 when consumed:
  %%
        Ref: O4
       Name: potion of healing
      Alias: POTION +POTION:HEALING
     Health: CURRENT→30

  This is a small, bright red potion.
  %%

  // A poison, decreases current health by -10 when consumed:
  %%
        Ref: O5
       Name: a potion
      Alias: POTION
     Health: CURRENT→-30

  This is a small, sickly green potion.
  %%

Note that these items, if created, will not currently work as intended.

Also note that we do not hint to the player what the effects of the
rings are, and only hint a little as to what the poison might do ;)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>