cmd: Drop redundant NPC|Player check for TAKE command
In the state.Take method there are two switchs. The first deals with
where and the second deals with what. However, there is a stray check in
the second switch checking for "where.Is&(NPC|Player) != 0" a second
time - it is already in the first switch.

This commit removes the redundant second check in the second switch.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>