core,docs: Implement minimum 1 second for ACTION/CLEANUP/RESET
Currently if AFTER and JITTER are not specified or are both zero for
ACTION, CLEANUP and RESET fields then the events do not occure. However,
the documented behaviour is that if ACTION, CLEANUP or RESET are defined
but AFTER and JITTER are not specified the minimum period for the event
should be one second.

This commit implements the minimum one second period with a modified
check. Instead of checking for ACTION+JITTER == 0 the check is now
ACTION+JITTER+DUE_IN == 0. This allows for a suspended event specified
by DUE_IN, without also registering a repeating event with AFTER and
JITTER.

The documentation in docs/zone-files.txt has been updated for details of
the new check.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>