core: In Init, Schedule HealthEvent if NPC needs healing
If an NPC is loaded with current health less than maximum health a
health event is not scheduled. If an NPC is holdable with current health
less than maximum health is junked a health event is not scheduled when
the mobile resets.

This commit will schedule a health event when Thing.Init is called for
an NPC that needs healing. This fixes both of the above issues.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>