core: Add PromptStyle asKey to types
This commit adds the new asKey constant PromptStyle to types. Initially
it is intended to be used to store the player's prompt type, such as
"BRIEF", "SHORT", "LONG", etc.

In the future it may also hold a player's chosen custom format. This is
why it is an asKey and not a intKey.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>