core: Add Script and preParse methods + modify parse for state
This commit adds a new Script method to compliment the current Parse
method. The common code from Parse has been moved into a preParse
method.

This is an interim step in restricting scripting commands so they cannot
be directly executed by players. However, players still need to be able
to execute scripting commands indirectly.

The Script method provides access to scripting commands from outside the
core package, as does the Parse method. For example the client package
can execute $POOF and $QUIT commands.

At the moment the only difference between Parse and Script is the flag
passed to preParse. A future commit will pass the flag on to the parse
method to implement checking of the commands entered by players.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>