core: Disallow players from directly executing scripting commands
This commit modifies the state.parse method to take an additional
allowScripting parameter. This controls whether scripting commands,
those prefixed with a dollar '$', can be executed or not. If a player
tried to directly execute, for example $HEALTH, they will receive the
reply "Eh?" - as for an unknown command.

Call sites for parse in state and Thing have been updated to pass the
additional parameter.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>