Makefile: Add build-chroot target (minimal chroot environment)
This commit adds a new build-chroot target for building a minimal
environment suitable for running 'as is' in a chroot:

  chroot/
  |-- bin
  |   `-- server
  `-- data
      |-- config.wrj
      |-- players
      `-- zones
          |-- reset.wrj
          |-- zinara_caves.wrj
          |-- zinara_south.wrj
          `-- zinara.wrj

This is mainly useful for isolated testing.

Example usage:

  >cd chroot
  >sudo chroot --userspec=wolfmud:wolfmud . /bin/server > ./server.log 2>&1 &
  [sudo] password for diddymus:
  >

Note: Make sure permissions are set appropriately for the user and group
chosen and that the data/players directory can be written to.

Note: Any set environment variables are carried into the chroot jail.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>