core: Thing.dump show event dueAt in UTC
Currently the event dueAt timestamp is shown in the local time of the
server:

  |- Events - len: 1
  |  `- Health: 0xc00005a690, at: Apr 13 19:13:21, in: 695ms

This can cause confusion if the server is in one timezone and you are
logged in from another timezone and using #DUMP for debugging. You might
not even know the timezone the server is in.

All logging by the server is already in UTC. It makes sense to use UTC
for the event dueAt as well.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>