core: add command history and recall for up to 3 commands
This commit adds a command history of up to the last 3 commands the
player has issued, a means of listing the last three commands and a
means of repeating any of the last three commands.

The last 3 commands can be listed using either '/HISTORY' or the
abbreviation '/!'. These commands are not themselves appended to the
history. If the player issues a command that is a duplicate of the most
recent command in the history then the duplicate is not appended to the
history.

Any of the last 3 commands can be repeated using '!', '!!' and '!!!'
respectively. With '!' the most recent command, '!!' the command before
that and '!!!' being the oldest command in the history. A repeated
command is not appended to the history.

The current history implementation stores the history in the state
rather than the thing, possibly as Thing.Any[History]. This allows the
history manipulation to be done outside of the critical path and outside
of taking the BWL.

Example:

  :
  :
  You see a small doll here.

  You see exits: west
  H:30/30>EXAMINE DOLL
  You examine the small doll.
  This is a small child's toy doll. They are wearing a tiny dress and a tiny hat.
  H:30/30>GET DOLL
  You get the small doll.
  H:30/30>DROP DOLL
  You drop the small doll.
  H:30/30>/HISTORY
  Your three most recent commands:
      !: DROP DOLL
     !!: GET DOLL
    !!!: EXAMINE DOLL
  H:30/30>!!!
  You examine the small doll.
  This is a small child's toy doll. They are wearing a tiny dress and a tiny hat.
  H:30/30>!!
  You get the small doll.
  H:30/30>!
  You drop the small doll.
  H:30/30>!!!
  You examine the small doll.
  This is a small child's toy doll. They are wearing a tiny dress and a tiny hat.
  H:30/30>/HISTORY
  Your three most recent commands:
      !: drop doll
     !!: get doll
    !!!: exam doll
  H:30/30>

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>