client: Limit input data from client (potential DOS attack mitigation)
Currently no limits are put on the length of a command a player can send
to the server. As the input is stored in a []byte buffer, a malicious
client could cause the buffer to consume a lot of the memory. This could
be used as a DOS attack - this has not been seen in the wild and has not
actually been exploitable during testing.

This commit limits the size of the input accepted by the server. If a
client sends a command longer than the limit to the server then all of
the input is discarded and the message "You type too much!" is sent to
the client. The current limit is 80 bytes.

The input buffer is now stored in the Client struct and shared between
the game client and the frontend. The input limit, and buffer size, is
set via the constant inputBufferSize. The client.input method has been
renamed client.read, which avoids conflicts and reads better.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>