core,frontend: Log when a player quits, willingly or forced
This commit logs when a player quits from the server, either willingly
or forced due to errors. This complements the logging of the player
login:

  [#UID-205] Login by: 90d9988c2b7014d622a62681e5643674
  [#UID-205] Quitting: 90d9988c2b7014d622a62681e5643674

Note, the login message has changed from "Login for:" to "Login by:" so
that the messages line up in the log and make scanning the log by eye
easier.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>