core: Add initial, basic combat event handler
This is a combat handler, initially implementing only the minimal
basics. There is a 50/50 chance who will win a combat round and damage
is a fixed 2+[1,2]. Rounds are 2 seconds apart.

There is a current limitation...

When the fight ends it is important that the killed opponent cannot
immediately respawn at the fight location. If this happens, and there is
more than one attacker, the opponent will be killed, respawn and
immediately be attacked by the remaining attackers. One way to avoid
this happening to players is to veto combat at starting locations.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>