core: Thing.suspend return false on failure to suspend
Currently Thing.suspend only returns false if the event is not
scheduled. However, it is possible that the timer could not be stopped
because it had already fired. In which case we still fallthrough and
clean up the event state.

If the timer cannot be stopped, because it had already fired, then false
should be returned after cleaning up the event state.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>