core: Thing.schedule only use dueIn > 0
Currently the check in Thing.schedule for using an event's dueIn value
is "dueIn != 0". This allows events to be scheduled in the past when
dueIn is negative. However, dueIn should always be zero or positive.

This commit changes the check for dueIn to be "dueIn > 0" so that only
future events are scheduled when dueIn is used.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>