core: add new minimal prompt style
This commit adds a new minimal prompt style "Hc␠". This style displays
'H', followed by the current health, followed by a single space. For
example:

  H30␠SNEEZE
  You sneeze.
  H30␠

Note that the '␠' will be rendered as a normal space to the user.

This is useful for screen readers where the user does not want to keep
hearing 'greater than' being read to them all the time, but want to know
their health.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>