core: only reschedule Action events if actor not in combat
If an Action event schedules a Combat event then further Action events
should not be scheduled until the combat finishes. However, it is not
currently possible for the Combat event to cancel the Action event
itself due to a) the Action event is already scheduled and firing b) the
Combat event launches within (a).

Currently the only was to achieve this is to save the Action event
settings somewhere, clear the current settings (to avoid the Action
event being re-scheduled), then restoring the settings when combat ends.
Very messy.

Instead a simpler approach has been take, to not re-schedule the Action
if in combat. This will need revisiting as it is not a generic solution
and only applies to Combat events.

An experiment has been tried that allows an event to be frozen. This is
different to suspending/cancelling an event. A frozen event is suspended
if scheduled, but cannot be scheduled/re-scheduled until thawed. This
worked quite well with the Combat event freezing the Action event when
combat starts and thawing the Action event when combat ended. However,
the implementation was quite invasive and the decision was made to go
with this simpler solution until after combat is implemented.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>