core: reset mobiles/NPCs at full health
When a mobile/NPC is reset it should have maximum current health.
However, if current health is specified in the zone file, the mobile/NPC
should still be initially loaded with the specified current health.

This change prevents a mobile that has been killed from immediatly being
killed again when it resets.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>