core: Update QUIT to end combat
When a player quits they could be in combat. In which case the combat
needs to be ended to clean up any opponents and their references. We
can't stop the player quitting as it could be due to a network errors or
something else. In the future it would be nice to allow the fight to
complete and then let the player quit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>