client,state: get+store terminal size and generate+enable sequences
This commit gets and stores the terminal size if available. If the size
is not available a default of 80x25 will be used. The sequences for
updating the output, status and input areas are generated once and
stored. The messenger method has been updated to emit the sequences to
update the output area and return to the input area.

This commit turns off the mailbox suffix so that the old prompt is not
written into the output. The state.Msg method has been updated to always
output on a new line as we no longer need to worry about being an actor
or not when writing messages.

This produces a working terminal display with an output area at the top,
a status line and an input area at the bottom. It should be noted that
the terminal should not be resized once a session has been started :(

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>