core: Update state.Parse to transfer input to output area
This commit updates state.Parse so that the player's input is written to
the output before displaying the results of their actions. This acts as
a point of reference for the player when reading the output. Without
this the output just becomes a sea of text and it is very hard to see
what just happened.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>