cmd: update botrunner for new terminal login
This commit updates the botrunner for the new login sequence expected
from the server due to the terminal changes. The recv method has been
updated to handle the new data flow of the login: we could receive "A"
then "B" as two network reads, or "AB" as a single network read. When
matching the incoming data recv now performs a simpler "does the input
contain the data we are looking for?" instead of matching prefixes and
suffixes.

As the new terminal handling generates more data, and there is a minor
pause during login, the delay between loading batches of 2048 bots has
been increased. This gives the server more time to 'catch up' but takes
longer to load up 32,000 or 64,000 bots :(

In addition the minimum number of steps a bot will perform before
quitting has been increased. This prevents bots starting to quit while
others are still logging in for large numbers of bots.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>