term: add Reset to return terminal to working state
This commit adds a Reset method to generate an escape sequence to reset
the scroll area to the size of the whole terminal, position the cursor
on the bottom line and reset the current colours. This sequence should
be used to return the terminal to a resonable working state.

Note: It is not acceptable to use the RIS (restore initial state)
sequence 'ESC c' as this also clears the screen. If the player is
dropped due to errors or quits and has to drop unique items they will
not be able to see the messages.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>