core: fix body parts lost when quitting with used unique items
Currently if a play quits while holding, wearing or wielding a unique
the item is forceably junked. However, the body slots are not allocated
back to the player. This results in the player having missing body slots
when they log back in.

This commit checks if items are in use before junking. If an item is
used the body slots are allocated back to the player. It should be noted
that we cannot sub-parse the REMOVE command to remove the items as this
may fail - for example if the item vetoes the REMOVE.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>