core: fix PUT/TAKE command breakage due to terminal changes
For the terminal changes the player's command is now transferred to the
output area by writing the command into the actor's buffer. This is now
causing the PUT and TAKE commands to fail in odd ways. This is due to
the 's.buf[s.actor] != nil' check now always being true.

This commit adds proper checking of the actor's buffer - as used by
Look. The proper check is to record the buffer's length and then to see
if the length has changed relative to the recorded length.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>