core: Add Armour+Damage to Thing Marshal/Unmarshal and preferredOrdering
This commit adds handling to Thing.Marshal and Thing.Unmarshal for the
Armour and Damage attributes. Note that values can be negative, giving a
penality, and so Marshal checks for != 0 rather than the normal == 0.

Both Armour and Damage have been added to the preferredOrdering of
attributes. They have been placed after health as body related
information.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>