client: upgrade/create players to have natural armour and damage
The standard humanoid model is health 30, armour 10 and damage 2+[0-2].
Other mobile should scale from this appropriately.

For items the standard is a dagger/short sword with damage 1+[0-3].

A standard for armour has not been decided yet. Probably +1 light
(cloth), +2 medium (leather), +3 heavy (ring/plate) per piece of armour,
with an addition +1 for large items such as jerkins and breastplates.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>