core: stop status updates disrupting message de-spamming
Currently messages sent to update the player's status line are sent as
non-priority. However, this means they are seen by mailbox as the last
message sent. This can lead to the following situation:

  You head talking nearby.
  <status update>
  You head talking nearby.

As the status update message is not seen in the output area this looks
like spamming to the player:

  You head talking nearby.
  You head talking nearby.

This commit updates StatusUpdate to send status updates as priority
messages. This prevents the apparent spamming. The status updates should
be priority anyway as we don't want status updates to be de-spammed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>