core: log who starts/ends combat
Who starts and who ends combat can be a big deal for players, causing
all sorts of arguments and aggravation.

This commit logs who starts a fight and with whom. It also logs who
survives and who gets killed:

  [#UID-192] a giant fungus slug attacked a cowardly imp (#UID-193)
  [#UID-192] a giant fungus slug killed a cowardly imp (#UID-193)

Currently all combat is logged. When the combat system is more mature,
logging for mobile vs mobile can probably be dropped.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>