recordjar: add encoding and decoding of double integers
This commit adds the encoding and decoding of two integers. In text form
the integers are encoded as "siSI" where 's' is the sign of 'i',
'i' is the first integer, 'S' is the sign of 'I' and 'I' is the second
integer. If 'I' is zero 'S' and 'I' may be ommited. If 'i' is positive
's' may be omitted.

For example:

  ""     =  0,  0
  "1"    =  1,  0
  "-1"   = -1,  0
  "+1"   =  1,  0
  "0+2"  =  0,  2
  "1+0"  =  1,  0
  "1+2"  =  1,  2
  "1-2"  =  1, -2
  "-1-2" = -1, -2

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>