core: add OnCombat to Thing Marshal/Unmarshal and preferredOrdering
This commit adds handling to Thing.Marshal and Thing.Unmarshal for the
OnCombat attribute. OnCombat is used to specify descriptive actions for
a successful combat attack. OnCombat can be used on players and mobiles
for natural attacks or on items for item specific attacks. In the future
OnCombat will be usable for skill/spell attacks as well.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>