docs: fix Go minimum version
Fix the Go minimum version in docs/compiling-from-source.txt and in the
go.mod file.

Minimum version should be 1.18 due to bytes.Cut and strings.Cut usage.
If the usage of Cut was replaced the minimum version would be Go 1.17
due to usage of math.MaxInt - however Go 1.17 is no longer supported.

Many thanks to user wade for spotting this and reporting it! :)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>