core,data: move bulk of help processing to load time, add groups
This commit removes most of the help heavy lifting out of the HELP
command processing and into the help build time processing. When the
library is loaded a new format method prepares the help pages ready for
being displayed to a user - instead of repeatedly formatting the pages
at runtime.

Pre-formatting the help pages also means that a more lightweight
structure can be used for the help library. Once pages are formatted
most of the build information can be discarded.

In addition help pages can now be grouped together using the GROUPS
field in a help topic. For example:

  %%
     Topic: MAIN
  Synopsis: Main help categories

  The following main help categories are available:
  %%
     Topic: COMMANDS
     Group: MAIN TOPICS
  Synopsis: Commands grouped by category

  The following categories of commands are available:
  %%
     Topic: ALL
     Group: MAIN TOPICS
  Synopsis: Help for all commands
  %%
     Topic: TOPICS
     Group: MAIN
  Synopsis: All help topics
  %%

Here the topics COMMANDS, ALL and TOPICS are grouped under 'MAIN'. As
for COMMANDS and ALL, a topic may appear in multiple groups. Groups may
also contain other groups providing a hierarchy of help topics. The top
level group is always called MAIN and there should always be a help
topics of MAIN present in the data/help.wrj file.

If help for 'MAIN' is requested, or a topic is not specified, the
following will be displayed:

  >/?
  Topic: MAIN
         Main help categories

  The following main help categories are available:

     COMMANDS  Commands grouped by category
          ALL  Help for all commands
       TOPICS  All help topics
  >

When a player consults the help pages using either /HELP or /? observers
in the same location will now see the message:

  Diddymus seems to be looking for answers...

Additional topics and groups have been added to data/help.wrj for
testing. This file needs to be re-written to provide proper help for
players on all topics and commands.

The help system can be disabled by setting "DISABLED: true" in the help
file - as per disabling zone files. In this case the players will see:

  >/HELP
  Sorry, help is currently unavailable.
  >

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>